fixup! fixup! Create custom replacer for circular json
