+++ src/initializers/log4js/json-appender.ts
@@ -1,4 +1,5 @@
 import * as lodash from 'lodash';
+import safeStringify from '../../helpers/jsonUtils';
 
 const jsonAppender = layout => {
   return logEvent => {
@@ -7,27 +8,6 @@ const jsonAppender = layout => {
     const json = safeStringify(event);
     console.log(json);
   };
-};
-
-const safeStringify = event => {
-  try {
-    return JSON.stringify(event);
-  } catch (error) {
-    let seen = new WeakSet();
-
-    const circularJson = JSON.stringify(event, (key, value) => {
-      if (typeof value === 'object' && value !== null) {
-        if (seen.has(value)) {
-          return 'circular_ref';
-        }
-        seen.add(value);
-      }
-      return value;
-    });
-
-    seen = null;
-    return circularJson;
-  }
 };
 
 export const createErrorLog = (layout, logEvent) => {
